python/servers/weather-forecast/pyproject.toml
[project]
name = "weather-forecast-server"
version = "0.1.0"
description = "기상청 단기예보 MCP 서버"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "data-go-kr-shared",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
weather-forecast-server = "src.server:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
data-go-kr-shared = { path = "../../shared" }
